+++ etna/analysis/outliers/median_outliers.py
@@ -4,11 +4,12 @@ import typing
 import numpy as np
 import pandas as pd
 
-from etna.datasets.tsdataset import TSDataset
+if typing.TYPE_CHECKING:
+    from etna.datasets import TSDataset
 
 
 def get_anomalies_median(
-    ts: TSDataset, window_size: int = 10, alpha: float = 3
+    ts: "TSDataset", window_size: int = 10, alpha: float = 3
 ) -> typing.Dict[str, typing.List[pd.Timestamp]]:
     """
     Get point outliers in time series using median model (estimation model-based method).