+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.1.0] - 2021-09-23
 ### Added 
 - MedianOutliersTransform, DensityOutliersTransform ([#30](https://github.com/tinkoff-ai/etna-ts/pull/30))
 - Issues and Pull Request templates
@@ -13,9 +13,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Pytorch-Forecasting models ([#29](https://github.com/tinkoff-ai/etna-ts/pull/29))
 - SARIMAX model ([#10](https://github.com/tinkoff-ai/etna-ts/pull/10))
 - Logging, including ConsoleLogger ([#46](https://github.com/tinkoff-ai/etna-ts/pull/46))
+- WandbLogger ([#71](https://github.com/tinkoff-ai/etna-ts/pull/71))
+- Correlation heatmap plotter ([#77](https://github.com/tinkoff-ai/etna-ts/pull/77))
 
 ### Changed
 - Backtest is fully parallel 
+- New default hyperparameters for CatBoost
 
 ### Fixed
 - Documentation fixes ([#55](https://github.com/tinkoff-ai/etna-ts/pull/55), [#53](https://github.com/tinkoff-ai/etna-ts/pull/53), [#52](https://github.com/tinkoff-ai/etna-ts/pull/52))
@@ -23,6 +26,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Regressors does not have enough history bug ([#35](https://github.com/tinkoff-ai/etna-ts/pull/35))
 - make_future(1) and make_future(2) bug
 - Fix working with 'cap' and 'floor' features in Prophet model ([#62](https://github.com/tinkoff-ai/etna-ts/pull/62)))
+- Fix saving init params for SARIMAXModel ([#81](https://github.com/tinkoff-ai/etna-ts/pull/81))
+- Imports of nn models, PytorchForecastingTransform and Transform ([#80](https://github.com/tinkoff-ai/etna-ts/pull/80)))
 
 ## [1.0.0] - 2021-09-05
 ### Added